refactor(shard): 扩展玩法 coord 迁出核心包
