+++ pallas/extensions/__init__.py
@@ -1,0 +1,1 @@
+"""Pallas 扩展玩法实现（非核心 shard coord 框架）。"""
