+++ pallas/extensions/coord/__init__.py
@@ -1,0 +1,1 @@
+"""扩展玩法分片协调实现。"""
