+++ pallas/extensions/coord/duel/__init__.py
@@ -1,0 +1,1 @@
+"""决斗扩展分片协调。"""
