+++ pallas/extensions/coord/maa/__init__.py
@@ -1,0 +1,1 @@
+"""MAA 扩展分片协调。"""
